issue #55
