+++ src/main/webapp/WEB-INF/spring-security.xml
@@ -12,7 +12,7 @@
     <http auto-config="true" use-expressions="true">
         <intercept-url pattern="/app/producer/dashboard/**" access="hasRole('ROLE_ADMIN')" />
         <intercept-url pattern="/app/photographers/shop/**" access="hasRole('ROLE_PHOTOGRAPHER')" />
-        <!--<intercept-url pattern="/app/customers/login/**" access="hasRole('ROLE_CUSTOMER')" />-->
+        <intercept-url pattern="/app/customers/login/**" access="hasRole('ROLE_CUSTOMER')" />
         <logout logout-url="/logout" logout-success-url="/app/logout"/>
         <csrf disabled="true"/>
         <form-login default-target-url="/app/"/>
