+++ src/main/webapp/WEB-INF/views/customers/home/index.twig
@@ -50,7 +50,7 @@
 
 
                     {% if username is empty %}
-                        <li><a href="/app/customers/login/new"><i class="fa fa-sign-in"></i> {{ labels.home_signup }}</a></li>
+                        <li><a href="/app/customers/register"><i class="fa fa-sign-in"></i> {{ labels.home_signup }}</a></li>
 
                         <li><a href="/app/customers/login"><i class="fa fa-user"></i> {{ labels.home_login }}</a></li>
